refactor: re-organize files into `utils` and `data_struct` for clarity
